+++ SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialActorChannel.cpp
@@ -20,6 +20,7 @@
 #include "Interop/GlobalStateManager.h"
 #include "Interop/SpatialReceiver.h"
 #include "Interop/SpatialSender.h"
+#include "Schema/AlwaysRelevant.h"
 #include "Schema/ClientRPCEndpoint.h"
 #include "Schema/ServerRPCEndpoint.h"
 #include "SpatialConstants.h"
@@ -38,11 +39,7 @@ namespace
 // This is a bookkeeping function that is similar to the one in RepLayout.cpp, modified for our needs (e.g. no NaKs)
 // We can't use the one in RepLayout.cpp because it's private and it cannot account for our approach.
 // In this function, we poll for any changes in Unreal properties compared to the last time we replicated this actor.
-#if ENGINE_MINOR_VERSION <= 20
-void UpdateChangelistHistory(TSharedPtr<FRepState>& RepState)
-#else
 void UpdateChangelistHistory(TUniquePtr<FRepState>& RepState)
-#endif
 {
 	check(RepState->HistoryEnd >= RepState->HistoryStart);
 
@@ -86,27 +83,32 @@ USpatialActorChannel::USpatialActorChannel(const FObjectInitializer& ObjectIniti
 {
 }
 
-#if ENGINE_MINOR_VERSION <= 20
-void USpatialActorChannel::Init(UNetConnection* InConnection, int32 ChannelIndex, bool bOpenedLocally)
-{
-	Super::Init(InConnection, ChannelIndex, bOpenedLocally);
-
-	NetDriver = Cast<USpatialNetDriver>(Connection->Driver);
-	check(NetDriver);
-	Sender = NetDriver->Sender;
-	Receiver = NetDriver->Receiver;
-}
-#else
 void USpatialActorChannel::Init(UNetConnection* InConnection, int32 ChannelIndex, EChannelCreateFlags CreateFlag)
 {
 	Super::Init(InConnection, ChannelIndex, CreateFlag);
 
+	// Actor Channels are pooled, so we must initialize internal state here.
+	bCreatedEntity = false;
+	bCreatingNewEntity = false;
+	EntityId = SpatialConstants::INVALID_ENTITY_ID;
+	bInterestDirty = false;
+	bNetOwned = false;
+	LastPositionSinceUpdate = FVector::ZeroVector;
+	TimeWhenPositionLastUpdated = 0.0f;
+
+	PendingDynamicSubobjects.Empty();
+	SavedOwnerWorkerAttribute.Empty();
+
+	FramesTillDormancyAllowed = 0;
+
+	ActorHandoverShadowData = nullptr;
+	HandoverShadowDataMap.Empty();
+
 	NetDriver = Cast<USpatialNetDriver>(Connection->Driver);
 	check(NetDriver);
 	Sender = NetDriver->Sender;
 	Receiver = NetDriver->Receiver;
 }
-#endif
 
 void USpatialActorChannel::DeleteEntityIfAuthoritative()
 {
@@ -123,7 +125,7 @@ void USpatialActorChannel::DeleteEntityIfAuthoritative()
 	{
 		// Workaround to delay the delete entity request if tearing off.
 		// Task to improve this: https://improbableio.atlassian.net/browse/UNR-841
-		if (Actor->GetTearOff())
+		if (Actor != nullptr && Actor->GetTearOff())
 		{
 			NetDriver->DelayedSendDeleteEntityRequest(EntityId, 1.0f);
 			// Since the entity deletion is delayed, this creates a situation,
@@ -133,11 +135,9 @@ void USpatialActorChannel::DeleteEntityIfAuthoritative()
 		}
 		else
 		{
-			Sender->SendDeleteEntityRequest(EntityId);
+			Sender->RetireEntity(EntityId);
 		}
 	}
-
-	Receiver->CleanupDeletedEntity(EntityId);
 }
 
 bool USpatialActorChannel::IsSingletonEntity()
@@ -145,11 +145,7 @@ bool USpatialActorChannel::IsSingletonEntity()
 	return NetDriver->GlobalStateManager->IsSingletonEntity(EntityId);
 }
 
-#if ENGINE_MINOR_VERSION <= 20
-bool USpatialActorChannel::CleanUp(const bool bForDestroy)
-#else
 bool USpatialActorChannel::CleanUp(const bool bForDestroy, EChannelCloseReason CloseReason)
-#endif
 {
 #if WITH_EDITOR
 	if (NetDriver != nullptr)
@@ -158,7 +154,8 @@ bool USpatialActorChannel::CleanUp(const bool bForDestroy, EChannelCloseReason C
 
 		if (bDeleteDynamicEntities &&
 			NetDriver->IsServer() &&
-			NetDriver->GetActorChannelByEntityId(EntityId) != nullptr)
+			NetDriver->GetActorChannelByEntityId(EntityId) != nullptr &&
+			CloseReason != EChannelCloseReason::Dormancy)
 		{
 			// If we're a server worker, and the entity hasn't already been cleaned up, delete it on shutdown.
 			DeleteEntityIfAuthoritative();
@@ -166,29 +163,40 @@ bool USpatialActorChannel::CleanUp(const bool bForDestroy, EChannelCloseReason C
 	}
 #endif
 
-	// Must cleanup actor and subobjects before UActorChannel::Cleanup as it will clear CreateSubObjects
-	Receiver->CleanupDeletedEntity(EntityId);
+	if (CloseReason != EChannelCloseReason::Dormancy)
+	{
+		// Must cleanup actor and subobjects before UActorChannel::Cleanup as it will clear CreateSubObjects.
+		NetDriver->PackageMap->RemoveEntityActor(EntityId);
+	}
+	else
+	{
+		NetDriver->RegisterDormantEntityId(EntityId);
+	}
 
-#if ENGINE_MINOR_VERSION <= 20
-	return UActorChannel::CleanUp(bForDestroy);
-#else
+	NetDriver->RemoveActorChannel(EntityId);
+
 	return UActorChannel::CleanUp(bForDestroy, CloseReason);
-#endif
 }
 
-#if ENGINE_MINOR_VERSION <= 20
-int64 USpatialActorChannel::Close()
-{
-	DeleteEntityIfAuthoritative();
-	return Super::Close();
-}
-#else
 int64 USpatialActorChannel::Close(EChannelCloseReason Reason)
 {
-	DeleteEntityIfAuthoritative();
+	if (Reason != EChannelCloseReason::Dormancy)
+	{
+		DeleteEntityIfAuthoritative();
+		NetDriver->PackageMap->RemoveEntityActor(EntityId);
+	}
+	else
+	{
+		// Closed for dormancy reasons, ensure we update the component state of this entity.
+		const bool bMakeDormant = true;
+		NetDriver->RefreshActorDormancy(Actor, bMakeDormant);
+		NetDriver->RegisterDormantEntityId(EntityId);
+	}
+
+	NetDriver->RemoveActorChannel(EntityId);
+
 	return Super::Close(Reason);
 }
-#endif
 
 bool USpatialActorChannel::IsDynamicArrayHandle(UObject* Object, uint16 Handle)
 {
@@ -203,7 +211,7 @@ void USpatialActorChannel::UpdateShadowData()
 {
 	check(Actor);
 
-	// If this channel was responsible for creating the channel, we do not want to initialize our shadow data
+	// If this channel was responsible for creating the actor, we do not want to initialize our shadow data
 	// to the latest state since there could have been state that has changed between creation of the entity
 	// and gaining of authority. Revisit this with UNR-1034
 	// TODO: UNR-1029 - log when the shadow data differs from the current state of the Actor.
@@ -237,11 +245,7 @@ FRepChangeState USpatialActorChannel::CreateInitialRepChangeState(TWeakObjectPtr
 	checkf(Object != nullptr, TEXT("Attempted to create initial rep change state on an object which is null."));
 	checkf(!Object->IsPendingKill(), TEXT("Attempted to create initial rep change state on an object which is pending kill. This will fail to create a RepLayout: "), *Object->GetName());
 
-#if ENGINE_MINOR_VERSION <= 20
-	FObjectReplicator& Replicator = FindOrCreateReplicator(Object).Get();
-#else
 	FObjectReplicator& Replicator = FindOrCreateReplicator(Object.Get()).Get();
-#endif
 
 	TArray<uint16> InitialRepChanged;
 
@@ -303,8 +307,14 @@ int64 USpatialActorChannel::ReplicateActor()
 
 	const UWorld* const ActorWorld = Actor->GetWorld();
 
-	// Time how long it takes to replicate this particular actor
-	STAT(FScopeCycleCounterUObject FunctionScope(Actor));
+#if STATS
+	// Group specific actor class stats by parent native class, which is what vanilla Unreal does.
+	UClass* ParentNativeClass = GetParentNativeClass(Actor->GetClass());
+	SCOPE_CYCLE_UOBJECT(ParentNativeClass, ParentNativeClass);
+#endif
+
+	// Group actors by exact class, one level below parent native class.
+	SCOPE_CYCLE_UOBJECT(ReplicateActor, Actor);
 
 	// Create an outgoing bunch (to satisfy some of the functions below).
 	FOutBunch Bunch(this, 0);
@@ -351,8 +361,7 @@ int64 USpatialActorChannel::ReplicateActor()
 
 	// Epic does this at the net driver level, per connection. See UNetDriver::ServerReplicateActors().
 	// However, we have many player controllers sharing one connection, so we do it at the actor level before replication.
-	APlayerController* PlayerController = Cast<APlayerController>(Actor);
-	if (PlayerController)
+	if (APlayerController* PlayerController = Cast<APlayerController>(Actor))
 	{
 		PlayerController->SendClientAdjustment();
 	}
@@ -373,11 +382,7 @@ int64 USpatialActorChannel::ReplicateActor()
 	// Update the replicated property change list.
 	FRepChangelistState* ChangelistState = ActorReplicator->ChangelistMgr->GetRepChangelistState();
 	bool bWroteSomethingImportant = false;
-#if ENGINE_MINOR_VERSION <= 20
-	ActorReplicator->ChangelistMgr->Update(Actor, Connection->Driver->ReplicationFrame, ActorReplicator->RepState->LastCompareIndex, RepFlags, bForceCompareProperties);
-#else
 	ActorReplicator->ChangelistMgr->Update(ActorReplicator->RepState.Get(), Actor, Connection->Driver->ReplicationFrame, RepFlags, bForceCompareProperties);
-#endif
 
 	const int32 PossibleNewHistoryIndex = ActorReplicator->RepState->HistoryEnd % FRepState::MAX_CHANGE_HISTORY;
 	FRepChangedHistory& PossibleNewHistoryItem = ActorReplicator->RepState->ChangeHistory[PossibleNewHistoryIndex];
@@ -421,6 +426,7 @@ int64 USpatialActorChannel::ReplicateActor()
 			Actor->ReplicateSubobjects(this, &Bunch, &RepFlags);
 
 			Sender->SendCreateEntityRequest(this);
+			bCreatedEntity = true;
 
 			// Since we've tried to create this Actor in Spatial, we no longer have authority over the actor since it hasn't been delegated to us.
 			Actor->Role = ROLE_SimulatedProxy;
@@ -434,6 +440,7 @@ int64 USpatialActorChannel::ReplicateActor()
 		}
 
 		bWroteSomethingImportant = true;
+
 		if (RepChanged.Num() > 0)
 		{
 			ActorReplicator->RepState->HistoryEnd++;
@@ -442,7 +449,22 @@ int64 USpatialActorChannel::ReplicateActor()
 
 	UpdateChangelistHistory(ActorReplicator->RepState);
 
+	// This would indicate we need to flush our state before we could consider going dormant. In Spatial, this
+	// dormancy can occur immediately (because we don't require acking), which means that dormancy can be thrashed
+	// on and off if AActor::FlushNetDormancy is being called (possibly because replicated properties are being updated
+	// within blueprints which invokes this call). Give a few frames before allowing channel to go dormant.
+	if (ActorReplicator->bLastUpdateEmpty == 0)
+	{
+		FramesTillDormancyAllowed = 2;
+	}
+	else if (FramesTillDormancyAllowed > 0)
+	{
+		--FramesTillDormancyAllowed;
+	}
+
 	ActorReplicator->RepState->LastChangelistIndex = ChangelistState->HistoryEnd;
+	ActorReplicator->RepState->OpenAckedCalled = true;
+	ActorReplicator->bLastUpdateEmpty = 1;
 
 	if (bCreatingNewEntity)
 	{
@@ -482,11 +504,7 @@ int64 USpatialActorChannel::ReplicateActor()
 		// Look for deleted subobjects
 		for (auto RepComp = ReplicationMap.CreateIterator(); RepComp; ++RepComp)
 		{
-#if ENGINE_MINOR_VERSION <= 20
-			if (!RepComp.Key().IsValid())
-#else
 			if (!RepComp.Value()->GetWeakObjectPtr().IsValid())
-#endif
 			{
 				FUnrealObjectRef ObjectRef = NetDriver->PackageMap->GetUnrealObjectRefFromNetGUID(RepComp.Value().Get().ObjectNetGUID);
 				if (ObjectRef.IsValid())
@@ -604,14 +622,18 @@ bool USpatialActorChannel::ReplicateSubobject(UObject* Object, const FReplicatio
 {
 	SCOPE_CYCLE_COUNTER(STAT_SpatialActorChannelReplicateSubobject);
 
+#if STATS
+	// Break down the subobject timing stats by parent native class.
+	UClass* ParentNativeClass = GetParentNativeClass(Object->GetClass());
+	SCOPE_CYCLE_UOBJECT(ReplicateSubobjectParentClass, ParentNativeClass);
+#endif
+
+	// Further break down the subobject timing stats by class.
+	SCOPE_CYCLE_UOBJECT(ReplicateSubobjectSpecificClass, Object);
+
 	bool bCreatedReplicator = false;
 
-#if ENGINE_MINOR_VERSION <= 20
-	bCreatedReplicator = !ReplicationMap.Contains(Object);
-	FObjectReplicator& Replicator = FindOrCreateReplicator(Object).Get();
-#else
 	FObjectReplicator& Replicator = FindOrCreateReplicator(Object, &bCreatedReplicator).Get();
-#endif
 
 	// If we're creating an entity, don't try replicating 
 	if (bCreatingNewEntity)
@@ -634,11 +656,7 @@ bool USpatialActorChannel::ReplicateSubobject(UObject* Object, const FReplicatio
 	}
 
 	FRepChangelistState* ChangelistState = Replicator.ChangelistMgr->GetRepChangelistState();
-#if ENGINE_MINOR_VERSION <= 20
-	Replicator.ChangelistMgr->Update(Object, Replicator.Connection->Driver->ReplicationFrame, Replicator.RepState->LastCompareIndex, RepFlags, bForceCompareProperties);
-#else
 	Replicator.ChangelistMgr->Update(Replicator.RepState.Get(), Object, Replicator.Connection->Driver->ReplicationFrame, RepFlags, bForceCompareProperties);
-#endif
 
 	const int32 PossibleNewHistoryIndex = Replicator.RepState->HistoryEnd % FRepState::MAX_CHANGE_HISTORY;
 	FRepChangedHistory& PossibleNewHistoryItem = Replicator.RepState->ChangeHistory[PossibleNewHistoryIndex];
@@ -682,6 +700,8 @@ bool USpatialActorChannel::ReplicateSubobject(UObject* Object, const FReplicatio
 
 	UpdateChangelistHistory(Replicator.RepState);
 	Replicator.RepState->LastChangelistIndex = ChangelistState->HistoryEnd;
+	Replicator.RepState->OpenAckedCalled = true;
+	Replicator.bLastUpdateEmpty = 1;
 
 	return RepChanged.Num() > 0;
 }
@@ -690,6 +710,23 @@ bool USpatialActorChannel::ReplicateSubobject(UObject* Obj, FOutBunch& Bunch, co
 {
 	// Intentionally don't call Super::ReplicateSubobject() but rather call our custom version instead.
 	return ReplicateSubobject(Obj, RepFlags);
+}
+
+bool USpatialActorChannel::ReadyForDormancy(bool bSuppressLogs /*= false*/)
+{
+ 	// Check Receiver doesn't have any pending operations for this channel
+ 	if (Receiver->IsPendingOpsOnChannel(this))
+ 	{
+ 		return false;
+ 	}
+
+	// Hasn't been waiting for dormancy long enough allow dormancy, soft attempt to prevent dormancy thrashing
+	if (FramesTillDormancyAllowed > 0)
+	{
+		return false;
+	}
+
+	return Super::ReadyForDormancy(bSuppressLogs);
 }
 
 TMap<UObject*, const FClassInfo*> USpatialActorChannel::GetHandoverSubobjects()
@@ -805,6 +842,7 @@ void USpatialActorChannel::SetChannelActor(AActor* InActor)
 			PackageMap->RemovePendingCreationEntityId(EntityId);
 		}
 		NetDriver->AddActorChannel(EntityId, this);
+		NetDriver->UnregisterDormantEntityId(EntityId);
 	}
 
 	// Set up the shadow data for the handover properties. This is used later to compare the properties and send only changed ones.
@@ -850,31 +888,31 @@ bool USpatialActorChannel::TryResolveActor()
 	return true;
 }
 
-FObjectReplicator& USpatialActorChannel::PreReceiveSpatialUpdate(UObject* TargetObject)
+FObjectReplicator* USpatialActorChannel::PreReceiveSpatialUpdate(UObject* TargetObject)
 {
+	// If there is no NetGUID for this object, we will crash in FObjectReplicator::StartReplicating, so we verify this here.
 	FNetworkGUID ObjectNetGUID = Connection->Driver->GuidCache->GetOrAssignNetGUID(TargetObject);
-	check(!ObjectNetGUID.IsDefault() && ObjectNetGUID.IsValid());
+	if (ObjectNetGUID.IsDefault() || !ObjectNetGUID.IsValid())
+	{
+		// SpatialReceiver tried to resolve this object in the PackageMap, but it didn't propagate to GuidCache.
+		// This could happen if the UnrealObjectRef was already mapped to a different object that's been destroyed.
+		UE_LOG(LogSpatialActorChannel, Error, TEXT("PreReceiveSpatialUpdate: NetGUID is invalid! Object: %s"), *TargetObject->GetPathName());
+		return nullptr;
+	}
 
 	FObjectReplicator& Replicator = FindOrCreateReplicator(TargetObject).Get();
 	TargetObject->PreNetReceive();
 	Replicator.RepLayout->InitShadowData(Replicator.RepState->StaticBuffer, TargetObject->GetClass(), (uint8*)TargetObject);
 
-	return Replicator;
+	return &Replicator;
 }
 
 void USpatialActorChannel::PostReceiveSpatialUpdate(UObject* TargetObject, const TArray<UProperty*>& RepNotifies)
 {
-	FNetworkGUID ObjectNetGUID = Connection->Driver->GuidCache->GetOrAssignNetGUID(TargetObject);
-	check(!ObjectNetGUID.IsDefault() && ObjectNetGUID.IsValid())
-
 	FObjectReplicator& Replicator = FindOrCreateReplicator(TargetObject).Get();
 	TargetObject->PostNetReceive();
 
-#if ENGINE_MINOR_VERSION <= 20
-	Replicator.RepNotifies = RepNotifies;
-#else
 	Replicator.RepState->RepNotifies = RepNotifies;
-#endif
 
 	Replicator.CallRepNotifies(false);
 
@@ -896,21 +934,27 @@ void USpatialActorChannel::OnCreateEntityResponse(const Worker_CreateEntityRespo
 
 	if (Op.status_code != WORKER_STATUS_CODE_SUCCESS)
 	{
-		// UNR-630 - Temporary hack to avoid failure to create entities due to timeout on large maps
 		if (Op.status_code == WORKER_STATUS_CODE_TIMEOUT)
 		{
 			UE_LOG(LogSpatialActorChannel, Warning, TEXT("Failed to create entity for actor %s Reason: %s. Retrying..."), *Actor->GetName(), UTF8_TO_TCHAR(Op.message));
 			Sender->SendCreateEntityRequest(this);
 		}
-		else
+#if !UE_BUILD_SHIPPING
+		// Commands can timeout locally, but still be processed by the runtime. When this occurs, our follow up `CreateEntityRequest` will be
+		// superfluous, as the entity will already be created. If we detect that the entity is already in our view, reduce the message severity
+		else if (NetDriver->StaticComponentView->GetComponentData<SpatialGDK::Position>(GetEntityId()) == nullptr)
 		{
 			UE_LOG(LogSpatialActorChannel, Error, TEXT("Failed to create entity for actor %s: Reason: %s"), *Actor->GetName(), UTF8_TO_TCHAR(Op.message));
 		}
+		else
+		{
+			UE_LOG(LogSpatialActorChannel, Verbose, TEXT("Failed to create entity for actor %s, entity already in view : Reason: %s"), *Actor->GetName(), UTF8_TO_TCHAR(Op.message));
+		}
+#endif // !UE_BUILD_SHIPPING
 
 		return;
 	}
 
-	bCreatedEntity = true;
 	UE_LOG(LogSpatialActorChannel, Verbose, TEXT("Created entity (%lld) for: %s."), Op.entity_id, *Actor->GetName());
 }
 
@@ -989,7 +1033,7 @@ FVector USpatialActorChannel::GetActorSpatialPosition(AActor* InActor)
 		USceneComponent* PawnRootComponent = Controller->GetPawn()->GetRootComponent();
 		Location = PawnRootComponent ? PawnRootComponent->GetComponentLocation() : FVector::ZeroVector;
 	}
-	else if (InActor->GetOwner() != nullptr && InActor->GetIsReplicated())
+	else if (InActor->GetOwner() != nullptr && InActor->GetOwner()->GetIsReplicated())
 	{
 		return GetActorSpatialPosition(InActor->GetOwner());
 	}
